# packages/contracts/openapi/academic.json
{
  "components": {
    "schemas": {
...
      "ApiResponseClassGroupAttendanceStat": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ClassGroupAttendanceStat"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ApiResponseClassGroupResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/ClassGroupResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseCourseResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/CourseResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseLessonResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/LessonResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseListSubscriptionResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionResponse"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponsePageResultClassGroupResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultClassGroupResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponsePageResultCourseResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultCourseResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseSubscriptionResponse": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "$ref": "#/components/schemas/SubscriptionResponse"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ApiResponseVoid": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {},
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ClassGroupAttendanceStat": {
        "properties": {
          "absent": {
            "format": "int32",
            "type": "integer"
          },
          "classGroupId": {
            "type": "string"
          },
          "classGroupName": {
            "type": "string"
          },
          "leave": {
            "format": "int32",
            "type": "integer"
          },
          "present": {
            "format": "int32",
            "type": "integer"
          },
          "rate": {
            "type": "number"
          },
          "total": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ClassGroupRequest": {
        "properties": {
          "branchId": {
            "type": "string"
          },
          "capacity": {
            "format": "int32",
            "type": "integer"
          },
          "courseId": {
            "type": "string"
          },
          "headTeacherId": {
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "nestedAvailability": {
            "$ref": "#/components/schemas/NestedTeacherAvailabilityRequest"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "tagColor": {
            "pattern": "^$|^#[0-9A-Fa-f]{6}$",
            "type": "string"
          },
          "teacherAvailabilityId": {
            "type": "string"
          }
        },
        "required": [
          "branchId",
          "name"
        ],
        "type": "object"
      },
      "ClassGroupResponse": {
        "properties": {
          "branchId": {
            "type": "string"
          },
          "capacity": {
            "format": "int32",
            "type": "integer"
          },
          "courseId": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "currentCount": {
            "format": "int32",
            "type": "integer"
          },
          "dayOfWeek": {
            "format": "int32",
            "type": "integer"
          },
          "endMinute": {
            "format": "int32",
            "type": "integer"
          },
          "headTeacherId": {
            "type": "string"
          },
          "headTeacherName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startMinute": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "tagColor": {
            "type": "string"
          },
          "teacherAvailabilityId": {
            "type": "string"
          },
          "teacherName": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CourseRequest": {
        "properties": {
          "ageMax": {
            "format": "int32",
            "type": "integer"
          },
          "ageMin": {
            "format": "int32",
            "type": "integer"
          },
          "coverUrl": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "lessonMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "CourseResponse": {
        "properties": {
          "ageMax": {
            "format": "int32",
            "type": "integer"
          },
          "ageMin": {
            "format": "int32",
            "type": "integer"
          },
          "coverUrl": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lessonMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "LessonResponse": {
        "properties": {
          "branchId": {
            "type": "string"
          },
          "classGroupId": {
            "type": "string"
          },
          "classGroupName": {
            "type": "string"
          },
          "classRoomId": {
            "type": "string"
          },
          "classRoomName": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "courseName": {
            "type": "string"
          },
          "endAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "startAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "teacherName": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "NestedTeacherAvailabilityRequest": {
        "properties": {
          "branchId": {
            "type": "string"
          },
          "capacity": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "dayOfWeek": {
            "format": "int32",
            "maximum": 7,
            "minimum": 1,
            "type": "integer"
          },
          "defaultClassRoomId": {
            "type": "string"
          },
          "endMinute": {
            "format": "int32",
            "maximum": 1439,
            "minimum": 0,
            "type": "integer"
          },
          "note": {
            "type": "string"
          },
          "startMinute": {
            "format": "int32",
            "maximum": 1439,
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "teacherId": {
            "type": "string"
          },
          "validFrom": {
            "format": "date",
            "type": "string"
          },
          "validTo": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "branchId",
          "capacity",
          "dayOfWeek",
          "endMinute",
          "startMinute",
          "teacherId",
          "validFrom"
        ],
        "type": "object"
      },
      "PageResultClassGroupResponse": {
        "properties": {
          "page": {
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ClassGroupResponse"
            },
            "type": "array"
          },
          "size": {
            "type": "string"
          },
          "total": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PageResultCourseResponse": {
        "properties": {
          "page": {
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/CourseResponse"
            },
            "type": "array"
          },
          "size": {
            "type": "string"
          },
          "total": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SubscriptionRequest": {
        "properties": {
          "note": {
            "maxLength": 256,
            "minLength": 0,
            "type": "string"
          },
          "source": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "studentId": {
            "type": "string"
          },
          "teacherAvailabilityId": {
            "type": "string"
          },
          "validFrom": {
            "format": "date",
            "type": "string"
          },
          "validTo": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "studentId",
          "teacherAvailabilityId",
          "validFrom"
        ],
        "type": "object"
      },
      "SubscriptionResponse": {
        "properties": {
          "branchId": {
            "type": "string"
          },
          "dayOfWeek": {
            "format": "int32",
            "type": "integer"
          },
          "endMinute": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "startMinute": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "studentId": {
            "type": "string"
          },
          "studentName": {
            "type": "string"
          },
          "teacherAvailabilityId": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "teacherName": {
            "type": "string"
          },
          "validFrom": {
            "format": "date",
            "type": "string"
          },
          "validTo": {
            "format": "date",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/class-groups": {
      "get": {
        "operationId": "list_8",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "branchId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "courseId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultClassGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "class-group-controller"
        ]
      },
      "post": {
        "operationId": "create_10",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClassGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "class-group-controller"
        ]
      }
    },
    "/api/class-groups/{id}": {
      "delete": {
        "operationId": "delete_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "class-group-controller"
        ]
      },
      "get": {
        "operationId": "get_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClassGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "class-group-controller"
        ]
      },
      "put": {
        "operationId": "update_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClassGroupResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "class-group-controller"
        ]
      }
    },
...
    "/api/courses": {
      "get": {
        "operationId": "list_6",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultCourseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-controller"
        ]
      },
      "post": {
        "operationId": "create_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCourseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-controller"
        ]
      }
    },
    "/api/courses/{id}": {
      "delete": {
        "operationId": "delete_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-controller"
        ]
      },
      "get": {
        "operationId": "get_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCourseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-controller"
        ]
      },
      "put": {
        "operationId": "update_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCourseResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "course-controller"
        ]
      }
    },
...
    "/api/lessons/{id}": {
      "delete": {
        "operationId": "delete_7",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "lesson-controller"
        ]
      },
      "get": {
        "operationId": "get_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLessonResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "lesson-controller"
        ]
      }
    },
...
    "/api/stats/attendance/class-group/{id}": {
      "get": {
        "operationId": "classGroup",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClassGroupAttendanceStat"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "attendance-stat-controller"
        ]
      }
    },
...
    "/api/subscriptions": {
      "get": {
        "operationId": "list_1",
        "parameters": [
          {
            "in": "query",
            "name": "studentId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "teacherId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListSubscriptionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "lesson-subscription-controller"
        ]
      },
      "post": {
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSubscriptionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "lesson-subscription-controller"
        ]
      }
    },
    "/api/subscriptions/{id}": {
      "delete": {
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "lesson-subscription-controller"
        ]
      },
      "put": {
        "operationId": "update_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSubscriptionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "lesson-subscription-controller"
        ]
      }
    },
...
  }
}